pyproject.toml
[project]
name = "forrester-spanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.11.0",
    "fastapi>=0.135.1",
    "google-cloud-spanner>=3.63.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.67.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "s2sphere>=0.2.5",
    "uvicorn>=0.42.0",
]
